Made workaround for jar-locations an extension method
