+++ metaserver/src/main/scala/scala/meta/languageserver/ScalametaLanguageServer.scala
@@ -52,6 +52,8 @@ class ScalametaLanguageServer(
 )(implicit s: Scheduler)
     extends LanguageServer(lspIn, lspOut) {
   implicit val cwd: AbsolutePath = config.cwd
+  private val tempSourcesDir: AbsolutePath =
+    cwd.resolve("target").resolve("sources")
   val (fileSystemSemanticdbSubscriber, fileSystemSemanticdbsPublisher) =
     ScalametaLanguageServer.semanticdbStream(cwd)
   val (compilerConfigSubscriber, compilerConfigPublisher) =
@@ -252,7 +254,7 @@ class ScalametaLanguageServer(
       )
       data <- symbolIndexer.definitionData(symbol)
       location <- symbolIndexer.definitionLocation(data)
-    } yield symbolIndexer.nonJarLocation(location)
+    } yield location.toNonJar(tempSourcesDir)
     DefinitionResult(locations.toList)
   }
 
@@ -272,8 +274,6 @@ class ScalametaLanguageServer(
         context.includeDeclaration
       )
     } yield location
-    // TODO(alexey) should it return references to the jars where symbol is defined? or should such references be filtered out instead?
-    // } yield symbolIndexer.nonJarLocation(location)
     ReferencesResult(locations)
   }
